pyproject.toml
[tool.poetry]
name = "Ateneus Cooperatius"
version = "0.1.0"
description = ""
authors = ["Codi Cooperatiu <hola@codi.coop>"]

[tool.poetry.dependencies]
python = "^3.9.2"
Django = "^3.2"
psycopg2-binary = "^2.9.3"
# Mantenir weasyprint a v52 degut a: https://github.com/Kozea/WeasyPrint/issues/1384
weasyprint = "62.3"
django-constance = {extras = ["database"], version = "^2.9.0"}
boto3 = "^1.24.22"
easy-thumbnails = "^2.8.1"
django-storages = "^1.12.3"
django-grappelli = "^3.0.3"
gunicorn = "^20.1.0"
django-modelclone = "^0.7.1"
openpyxl = "^3.0.10"
django-object-actions = "^4.0.0"
django-maintenance-mode = "^0.16.3"
django-mail-queue = "^3.2.5"
django-logentry-admin = "^1.1.0"
django-tagulous = "^1.3.3"
django-extensions = "^3.1.5"
django-environ = "^0.9.0"
six = "^1.16.0"
whitenoise = "^6.2.0"
django-sendgrid-v5 = "^1.2.0"
django-summernote = "^0.8.20"
celery = "^5.2.7"
qrcode = "^7.4.2"
loggly-python-handler = "^1.0.1"
icecream = "^2.1.3"
django-localflavor = "^4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
